span `#root` to `op_span`
