+++ hydroflow/tests/compile-fail/surface_demuxenum_wrongenum.stderr
@@ -33,18 +33,8 @@ error[E0599]: no variant named `Square` found for enum `std::option::Option<()>`
 error[E0277]: the trait bound `std::option::Option<()>: DemuxEnum<_>` is not satisfied
   --> tests/compile-fail/surface_demuxenum_wrongenum.rs:17:15
    |
-12 |       let mut df = hydroflow_syntax! {
-   |  __________________-
-13 | |         my_demux = source_iter([
-14 | |             Shape::Rectangle { w: 10.0, h: 8.0 },
-15 | |             Shape::Square(9.0),
-16 | |             Shape::Circle { r: 5.0 },
-17 | |         ]) -> demux_enum::<Option<()>>();
-   | |               ^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `DemuxEnum<_>` is not implemented for `std::option::Option<()>`
-...  |
-20 | |         my_demux[Square] -> for_each(std::mem::drop);
-21 | |     };
-   | |_____- required by a bound introduced by this call
+17 |         ]) -> demux_enum::<Option<()>>();
+   |               ^^^^^^^^^^^^^^^^^^^^^^^^^^ the trait `DemuxEnum<_>` is not implemented for `std::option::Option<()>`
    |
    = note: Ensure there is exactly one output for each enum variant.
    = note: Ensure that the type for each output is a tuple of the field for the variant: `()`, `(a,)`, or `(a, b, ...)`.
