+++ hydroflow/tests/compile-fail/surface_source_interval_badarg.stderr
@@ -1,13 +1,11 @@
 error[E0308]: mismatched types
  --> tests/compile-fail/surface_source_interval_badarg.rs:7:25
   |
-5 |       let mut df = hydroflow_syntax! {
-  |  __________________-
-6 | |         // Should be a `Duration`.
-7 | |         source_interval(5) -> for_each(std::mem::drop);
-  | |                         ^ expected `Duration`, found integer
-8 | |     };
-  | |_____- arguments to this function are incorrect
+7 |         source_interval(5) -> for_each(std::mem::drop);
+  |         ----------------^-
+  |         |               |
+  |         |               expected `Duration`, found integer
+  |         arguments to this function are incorrect
   |
 note: function defined here
  --> $CARGO/tokio-1.35.1/src/time/interval.rs
